requirements.txt
streamlit==1.25.0
opencv-python-headless==4.8.0.76
tensorflow==2.13.0
transformers==4.31.0
Pillow==9.5.0
pytesseract==0.3.10
numpy==1.23.5


opencv-python
scikit-image
pillow
pixellib
tensorflow-gpu